refactor(ui): вынесена preview infrastructure SidebarView
preview host, scenario factory, seed data и preview-only helpers для Sidebar перенесены из SidebarView.swift в отдельный SidebarPreviewData
runtime SidebarView очищен от preview wiring и теперь ближе к screen composition pattern остальных экранов
